pyproject.toml
[project]
name = "pad2skills"
version = "0.1.0"
description = "Build a Python pipeline that converts World Bank Project Appraisal Document PDFs into structured datasets of extracted skills and occupations, enabling an interactive dashboard and chatbot for exploration by sector and project."
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "dash>=3.3.0",
    "docling>=2.66.0",
    "matplotlib>=3.10.8",
    "numpy<2.0.0",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "torch<2.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
